backend/requirements.txt
# Core Dependencies
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic_core==2.33.2
starlette==0.47.2

# AI and LLM Tools
google-generativeai==0.8.5
langchain-core>=0.2.22,<0.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.2.0
filetype==1.2.0
lxml==6.0.1

# Database
supabase==2.18.1
supabase_auth==2.12.3

# HTTP and Networking
requests==2.32.5
httpx==0.28.1

# Utilities
python-dotenv==1.1.1
PyYAML==6.0.2
orjson==3.11.2
cachetools==5.5.2
tenacity==9.1.2
tqdm==4.67.1
# API Framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2
python-multipart==0.0.20

# AI and LLM Tools
google-generativeai==0.8.5
langchain-core>=0.2.22,<0.3  # Core utilities for both LangGraph and LangChain
langgraph==0.1.15  # Added for workflow management

# Document Processing
PyPDF2==3.0.1
python-docx==1.2.0
filetype==1.2.0
lxml==6.0.1

# Database and Authentication
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
storage3==0.12.1
pydantic==2.11.7
pydantic_core==2.33.2

# HTTP and Networking
requests==2.32.5
httpx==0.28.1
email-validator==2.1.1
websockets==15.0.1

# Utilities
python-dotenv==1.1.1
PyYAML==6.0.2
orjson==3.11.2
cachetools==5.5.2
tenacity==9.1.2
tqdm==4.67.1

# Google API Dependencies
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
googleapis-common-protos==1.70.0

# Other Dependencies
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
hyperframe==6.1.0
idna==3.10
jsonpointer==3.0.0
packaging==25.0
postgrest==1.1.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
realtime==2.7.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
StrEnum==0.4.15
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
zstandard==0.24.0
zstandard==0.24.0
